requirements.txt
pydantic-2.9.1
streamlit-1.38.0